Version 0.1

+++ model/Profesor.php
@@ -5,7 +5,7 @@ class Profesor {
 	private $id;
 	private $nombre;
 	private $apellidos;
-  private $email;
+	private $email;
 
 	public function __construct( $id = null ){
 		if( $id ){
@@ -14,7 +14,7 @@ class Profesor {
 			$this->id = $profesor['id'];
 			$this->nombre = $profesor['nombre'];
 			$this->apellidos = $profesor['apellidos'];
-      $this->email = $profesor['email'];
+			$this->email = $profesor['email'];
 		}
 	}
 
@@ -47,7 +47,7 @@ class Profesor {
 			$params = [
 				'nombre' => $this->nombre,
 				'apellidos' => $this->apellidos,
-        'email' => $this->email,
+        		'email' => $this->email,
 			];
 			Database::insertProfesor( $params );
 			return $this->id = Database::insertId();
@@ -56,27 +56,32 @@ class Profesor {
 	}
 
 	public function update(){
-		$params = [
-			'nombre' => $this->nombre,
-			'apellidos' => $this->apellidos,
-      'email' => $this->email,
-		];
-		return Database::updateProfesor( $this->id, $params );
+		if( $this->id ){
+			$params = [
+				'nombre' => $this->nombre,
+				'apellidos' => $this->apellidos,
+	      		'email' => $this->email,
+			];
+			return Database::updateProfesor( $this->id, $params );
+		}
+		return false;
 	}
 
 	public function remove(){
-		return Database::deleteProfesor( $this->id );
+		if( $this->id )
+			return Database::deleteProfesor( $this->id );
+		return false;
 	}
 
 	public static function get(){
-		$profesor_db = Database::getProfesorList();
+		$profesores_db = Database::getProfesorList();
 		$profesor_arr = [];
-		while( $profesor_db = $profesor_db->fetch_assoc() ){
+		while( $profesor_db = $profesores_db->fetch_assoc() ){
 			$profesor = new Profesor();
 			$profesor->id( $profesor_db['id'] );
 			$profesor->nombre( $profesor_db['nombre'] );
 			$profesor->apellidos( $profesor_db['apellidos'] );
-      $profesor->email( $profesor_db['email'] );
+			$profesor->email( $profesor_db['email'] );
 			$profesor_arr[] = $profesor;
 		}
 		return $profesor_arr;
